Implemented saving of elements

+++ src/main/java/com/derekprovance/biometrics/biometricsapi/api/meals/MealRepository.java
@@ -3,5 +3,5 @@ package com.derekprovance.biometrics.biometricsapi.api.meals;
 import org.springframework.data.repository.CrudRepository;
 
 public interface MealRepository extends CrudRepository<MealEntry, Integer> {
-    MealEntry findByLogId(Integer logId);
+    MealEntry findByLogId(Long logId);
 }
